+++ modules/clickhouse/src/main/resources/META-INF/services/org.testcontainers.r2dbc.R2DBCDatabaseContainerProvider
@@ -1,1 +1,1 @@
-org.testcontainers.containers.ClickHouseR2DBCDatabaseContainerProvider
+org.testcontainers.clickhouse.ClickHouseR2DBCDatabaseContainerProvider
